third_party/blink/web_tests/http/tests/websocket/interleaved-fragments-expected.txt

Test whether WebSocket rejects interleaved fragmented frames.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS onerror() was called
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
